fix: deployDehydration Service
