+++ .env.self-hosted.example
@@ -56,8 +56,8 @@ NEXT_PUBLIC_CDN_URL=
 # ------------------------------------
 # AI Provider (at least one required)
 # ------------------------------------
-ANTHROPIC_API_KEY=
-# OPENAI_API_KEY=
+OPENAI_API_KEY=
+# ANTHROPIC_API_KEY=
 
 # ------------------------------------
 # OAuth (optional — for email integrations)
