+++ packages/lib/src/dehydration/types.ts
@@ -32,6 +32,7 @@ export interface DehydratedEnvironment {
   apiUrl: string
   homepageUrl: string
   docsUrl: string
+  devPortalUrl: string
   cdnUrl: string
   // External services
   stripe: {
@@ -80,6 +81,7 @@ export interface DehydratedUser {
   // Auth metadata
   providers: string[]
   hasPassword: boolean
+  isSuperAdmin: boolean
   registrationMethod: 'oauth' | 'email' | 'phone' | 'mixed'
   memberships: Array<{
     id: string
